# api/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/medias/{filename}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/groupPictures/{filename}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/avatars/{filename}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
...
  },
...
}